+++ packages/implementations/react/CHANGELOG.md
@@ -2,23 +2,29 @@
 
 `@rarui-react/components` components is a component library built with [React](https://reactjs.org).
 
+## 2024-04-10 `1.3.0`
+
+#### 🎉 New features
+
+- Added new atomic `Title` component. ([#18](https://git.rarolabs.com.br/frontend/rarui/-/merge_requests/18) by [@junior](https://git.rarolabs.com.br/junior))
+
 ## 2024-04-10 `1.2.0`
 
 #### 🎉 New features
 
-- Added new atomic `Box` component. ([#15](https://git.rarolabs.com.br/frontend/rarui/pull/15) by [@junior](https://git.rarolabs.com.br/junior))
+- Added new atomic `Box` component. ([#16](https://git.rarolabs.com.br/frontend/rarui/-/merge_requests/16) by [@junior](https://git.rarolabs.com.br/junior))
 
 ## 2024-03-23 `1.1.0`
 
 #### 🎉 New features
 
-- Added new atomic `Link` component. ([#8](https://git.rarolabs.com.br/frontend/rarui/pull/8) by [@junior](https://git.rarolabs.com.br/junior))
+- Added new atomic `Link` component. ([#8](https://git.rarolabs.com.br/frontend/rarui/-/merge_requests/8) by [@junior](https://git.rarolabs.com.br/junior))
 
 ## 2024-03-22 `1.0.0`
 
 #### 🎉 New features
 
-- Added new atomic `Button` component. ([#5](https://git.rarolabs.com.br/frontend/rarui/pull/5) by [@junior](https://git.rarolabs.com.br/junior))
+- Added new atomic `Button` component. ([#5](https://git.rarolabs.com.br/frontend/rarui/-/merge_requests/5) by [@junior](https://git.rarolabs.com.br/junior))
 
 <!-- #### 🛠 Breaking changes -->